refactor(BA-7780): file the app config allow list and fragment types with their ids
Both entity types sat in app_config.py while their EntityIdentifier lived in the
file named after the kind, which then imported the type back. Each kind now
declares its type and its id in one file, and app_config.py keeps only the
merged config the caller reads.
