+++ src/ai/backend/common/data/entity/app_config.py
@@ -4,35 +4,9 @@ from uuid import UUID
 from ai.backend.common.data.entity.types import EntityType
 
 __all__ = (
-    "AppConfigAllowListEntityType",
     "AppConfigEntityType",
-    "AppConfigFragmentEntityType",
     "AppConfigScopeID",
 )
-
-
-class AppConfigAllowListEntityType(EntityType):
-    @override
-    @classmethod
-    def name(cls) -> str:
-        return "app_config_allow_list"
-
-    @override
-    @classmethod
-    def description(cls) -> str:
-        return "The set of app config keys a scope may set."
-
-
-class AppConfigFragmentEntityType(EntityType):
-    @override
-    @classmethod
-    def name(cls) -> str:
-        return "app_config_fragment"
-
-    @override
-    @classmethod
-    def description(cls) -> str:
-        return "One key and value of an app config, set in one scope."
 
 
 # The merged config a caller reads; the fragments it is merged from are their own type.
