+++ src/ai/backend/common/data/entity/app_config_allow_list.py
@@ -1,9 +1,23 @@
 from typing import override
 
-from ai.backend.common.data.entity.app_config import AppConfigAllowListEntityType
 from ai.backend.common.data.entity.types import EntityIdentifier, EntityType
 
-__all__ = ("AppConfigAllowListID",)
+__all__ = (
+    "AppConfigAllowListEntityType",
+    "AppConfigAllowListID",
+)
+
+
+class AppConfigAllowListEntityType(EntityType):
+    @override
+    @classmethod
+    def name(cls) -> str:
+        return "app_config_allow_list"
+
+    @override
+    @classmethod
+    def description(cls) -> str:
+        return "The set of app config keys a scope may set."
 
 
 class AppConfigAllowListID(EntityIdentifier):
