+++ src/ai/backend/common/data/entity/app_config_fragment.py
@@ -1,9 +1,23 @@
 from typing import override
 
-from ai.backend.common.data.entity.app_config import AppConfigFragmentEntityType
 from ai.backend.common.data.entity.types import EntityIdentifier, EntityType
 
-__all__ = ("AppConfigFragmentID",)
+__all__ = (
+    "AppConfigFragmentEntityType",
+    "AppConfigFragmentID",
+)
+
+
+class AppConfigFragmentEntityType(EntityType):
+    @override
+    @classmethod
+    def name(cls) -> str:
+        return "app_config_fragment"
+
+    @override
+    @classmethod
+    def description(cls) -> str:
+        return "One key and value of an app config, set in one scope."
 
 
 class AppConfigFragmentID(EntityIdentifier):
