+++ src/ai/backend/manager/services/app_config/actions/fragment/bulk_upsert.py
@@ -4,7 +4,7 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import override
 
-from ai.backend.common.data.entity.app_config import AppConfigFragmentEntityType
+from ai.backend.common.data.entity.app_config_fragment import AppConfigFragmentEntityType
 from ai.backend.common.data.entity.types import (
     EntityIdentifier,
     EntityType,
