+++ src/ai/backend/manager/services/app_config/actions/fragment/global_bulk_upsert.py
@@ -4,7 +4,7 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 from typing import override
 
-from ai.backend.common.data.entity.app_config import AppConfigFragmentEntityType
+from ai.backend.common.data.entity.app_config_fragment import AppConfigFragmentEntityType
 from ai.backend.common.data.entity.types import EntityType
 from ai.backend.manager.actions.v2.ops.base import AtomicUpsertGlobalEntityOpsAction
 from ai.backend.manager.data.app_config.types import AppConfigFragmentData
